+++ adapters/Moloco/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Moloco iOS Mediation Adapter Changelog
 
+#### Next version
+- Implemented a native ad.
+
 #### [Version 3.7.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/moloco/MolocoAdapter-3.7.0.0.zip)
 - Verified compatibility with Moloco SDK version 3.7.0.